Move Command classes to views package
