+++ src/main/java/tundra/views/DeleteCommand.java
@@ -1,14 +1,14 @@
-package tundra.models;
+package tundra.views;
 
 import tundra.exceptions.TundraException;
+import tundra.models.Task;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
-import tundra.views.Ui;
 
-public class UnmarkCommand extends Command {
+public class DeleteCommand extends Command {
 
     @Override
-    public void init(String fullCommand) {
+    public void init (String fullCommand) {
         String[] parts = fullCommand.split(" ", 2);
         setArguments(parts);
     }
@@ -19,16 +19,17 @@ public class UnmarkCommand extends Command {
         try {
             int i = Integer.parseInt(arguments[1]) - 1;
             Task task = taskList.get(i);
-            task.setCompleted(false);
-            taskList.update(i, task);
+            taskList.remove(i);
             ui.printMessage(
-                    "Ok, I've marked this task as not done yet:\n",
-                    "\t" + task + "\n"
+                    "Noted. I've removed this task:\n",
+                    "\t" + task + "\n",
+                    "Now you have " + taskList.size() + " task(s) in the list.\n"
             );
         } catch (ArrayIndexOutOfBoundsException | NumberFormatException e) {
-            throw new TundraException("Incorrect syntax. Usage: mark [number]");
+            throw new TundraException("Incorrect syntax. Usage: delete [number]");
         } catch (IndexOutOfBoundsException e) {
             throw new TundraException("No such task. Enter 'list' to see all tasks");
         }
     }
+
 }
