+++ src/main/java/tundra/views/MarkCommand.java
@@ -1,14 +1,14 @@
-package tundra.models;
+package tundra.views;
 
 import tundra.exceptions.TundraException;
+import tundra.models.Task;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
-import tundra.views.Ui;
 
-public class DeleteCommand extends Command {
+public class MarkCommand extends Command {
 
     @Override
-    public void init (String fullCommand) {
+    public void init(String fullCommand) {
         String[] parts = fullCommand.split(" ", 2);
         setArguments(parts);
     }
@@ -19,14 +19,14 @@ public class DeleteCommand extends Command {
         try {
             int i = Integer.parseInt(arguments[1]) - 1;
             Task task = taskList.get(i);
-            taskList.remove(i);
+            task.setCompleted(true);
+            taskList.update(i, task);
             ui.printMessage(
-                    "Noted. I've removed this task:\n",
-                    "\t" + task + "\n",
-                    "Now you have " + taskList.size() + " task(s) in the list.\n"
+                    "Nice! I've marked this task as done:\n",
+                    "\t" + task + "\n"
             );
         } catch (ArrayIndexOutOfBoundsException | NumberFormatException e) {
-            throw new TundraException("Incorrect syntax. Usage: delete [number]");
+            throw new TundraException("Incorrect syntax. Usage: mark [number]");
         } catch (IndexOutOfBoundsException e) {
             throw new TundraException("No such task. Enter 'list' to see all tasks");
         }
