+++ src/main/java/tundra/views/UnmarkCommand.java
@@ -1,11 +1,11 @@
-package tundra.models;
+package tundra.views;
 
 import tundra.exceptions.TundraException;
+import tundra.models.Task;
 import tundra.utils.Storage;
 import tundra.utils.TaskList;
-import tundra.views.Ui;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
 
     @Override
     public void init(String fullCommand) {
@@ -19,10 +19,10 @@ public class MarkCommand extends Command {
         try {
             int i = Integer.parseInt(arguments[1]) - 1;
             Task task = taskList.get(i);
-            task.setCompleted(true);
+            task.setCompleted(false);
             taskList.update(i, task);
             ui.printMessage(
-                    "Nice! I've marked this task as done:\n",
+                    "Ok, I've marked this task as not done yet:\n",
                     "\t" + task + "\n"
             );
         } catch (ArrayIndexOutOfBoundsException | NumberFormatException e) {
@@ -31,5 +31,4 @@ public class MarkCommand extends Command {
             throw new TundraException("No such task. Enter 'list' to see all tasks");
         }
     }
-
 }
